+++ src/main/java/module-info.java
@@ -5,4 +5,7 @@ module com.example.menstrualender {
 
     opens com.example.menstrualender to javafx.fxml;
     exports com.example.menstrualender;
+    opens com.example.menstrualender.view to javafx.fxml;
+    exports com.example.menstrualender.view;
+
 }
