+++ src/main/resources/com/example/menstrualender/view/login.fxml
@@ -8,7 +8,7 @@
 
 <AnchorPane xmlns="http://javafx.com/javafx"
             xmlns:fx="http://javafx.com/fxml"
-            fx:controller="$CONTROLLER_NAME$"
+            fx:controller=""
             prefHeight="400.0" prefWidth="600.0" >
 
 </AnchorPane>
